firebase/functions/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudevents==1.11.0
cryptography==46.0.2
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
firebase-functions==0.4.3
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
functions-framework==3.9.2
geojson==3.2.0
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-events==0.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.8
langchain-community==0.3.30
langchain-core==1.1.0
langchain-openai==0.3.34
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==1.0.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.32
MarkupSafe==3.0.3
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==2.1.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
pyowm==3.5.0
PySocks==1.7.1
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvicorn-worker==0.4.0
watchdog==6.0.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.20.1
zstandard==0.25.0
